Move code into video component
