+++ src/app/components/Article.jsx
@@ -3,6 +3,7 @@ import Helmet from 'react-helmet';
 import 'isomorphic-fetch';
 import styled from 'styled-components';
 import Header from './Header';
+import Video from './Video';
 
 const Headline = styled.h1`
   color: #222;
@@ -34,56 +35,15 @@ class Article extends Component {
 
   render() {
     const { headline } = this.state;
-    const mediaPlayerStyles = {
-      height: 270,
-      width: 480,
-    };
 
     return (
       <Fragment>
         <Helmet htmlAttributes={{ lang: 'en-GB' }}>
           <title>{headline}</title>
-          <script
-            type="text/javascript"
-            src="https://static.bbci.co.uk/frameworks/requirejs/0.13.0/sharedmodules/require.js"
-          />
-          <script type="text/javascript">
-            {`
-            const bbcRequireMap = {
-              "bump-4": "//emp.bbci.co.uk/emp/bump-4/bump-4",
-            };
-            require({ paths: bbcRequireMap });
-          `}
-          </script>
         </Helmet>
         <Header />
         <Headline>{headline}</Headline>
-        <div id="mediaPlayer12345678" style={mediaPlayerStyles} />
-        <script type="text/javascript">
-          {`
-            const settings = {
-              product: 'news',
-              responsive: true,
-              playlistObject: {
-                title: 'Butterfly photobombs koala film shoot at Australia zoo',
-                holdingImageURL:
-                  'https://ichef.bbci.co.uk/images/ic/$recipe/p049srmr.jpg',
-                items: [
-                  {
-                    versionID: 'p049sq7k',
-                    kind: 'programme',
-                    duration: 37,
-                  },
-                ],
-              },
-            };
-
-            require(['bump-4'], function (bump) {
-              var mediaPlayer = bump.player(document.getElementById('mediaPlayer'), settings);
-              mediaPlayer.load();
-            });
-          `}
-        </script>
+        <Video />
       </Fragment>
     );
   }
