+++ src/app/components/Video/index.stories.jsx
@@ -1,0 +1,5 @@
+import React from 'react';
+import { storiesOf } from '@storybook/react'; // eslint-disable-line import/no-extraneous-dependencies
+import Video from './index';
+
+storiesOf('Video', module).add('default', () => <Video />);
